create context, add likes and removing

+++ src/utils/Api.js
@@ -49,16 +49,28 @@ class Api {
     });
   }
 
-  addLike(id) {
-    return this._fetch(`${this._url}cards/${id}/likes`, {
-      method: "PUT",
-    });
-  }
+  // addLike(id) {
+  //   return this._fetch(`${this._url}cards/${id}/likes`, {
+  //     method: "PUT",
+  //   });
+  // }
 
-  deleteLike(id) {
-    return this._fetch(`${this._url}cards/${id}/likes`, {
-      method: "DELETE",
-    });
+  // deleteLike(id) {
+  //   return this._fetch(`${this._url}cards/${id}/likes`, {
+  //     method: "DELETE",
+  //   });
+  // }
+
+  changeLikeCardStatus(id, isLiked) {
+    if (isLiked) {
+      return this._fetch(`${this._url}cards/${id}/likes`, {
+        method: "PUT",
+      });
+    } else {
+      return this._fetch(`${this._url}cards/${id}/likes`, {
+        method: "DELETE",
+      });
+    }
   }
 
   deleteCard(id) {
